pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "widemem-ai"
version = "1.3.0"
description = "Next-gen AI memory layer with hierarchical memory, importance scoring, and temporal retrieval"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.10"
authors = [{ name = "Radu Cioplea", email = "radu@cioplea.com" }]
keywords = ["ai", "memory", "llm", "embeddings", "vector-search"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "pydantic>=2.0",
    "openai>=1.0",
    "faiss-cpu>=1.7",
    "numpy>=1.24",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "ruff>=0.4",
]
anthropic = ["anthropic>=0.30"]
ollama = ["ollama>=0.2"]
qdrant = ["qdrant-client>=1.9"]
sentence-transformers = ["sentence-transformers>=2.0"]
all = [
    "anthropic>=0.30",
    "ollama>=0.2",
    "qdrant-client>=1.9",
    "sentence-transformers>=2.0",
]

[tool.hatch.build.targets.wheel]
packages = ["widemem"]

[project.urls]
Homepage = "https://widemem.ai"
Repository = "https://github.com/remete618/widemem-ai"

[tool.ruff]
target-version = "py310"
line-length = 140

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
